gateway/requirements.txt
pyzmq==15.2.0
websockets==3.1
